Refactor test structure and fix layer linking
Reorganize tests to mirror MLIR directory structure and fix the inlined
layer hack by using proper modular linking.

Test structure changes:
- Create tests/components/ and tests/layers/ to mirror components/ and layers/
- Move all component tests to tests/components/
- Move layer test to tests/layers/
- Move generator to tests/layers/transformer_moe_config.py (co-located with test)
- Remove layers/__init__.py - layers/ is now MLIR-only, not a Python package

Fixed layer linking (removed inlining hack):
- Generate separate accessor module (@model_params namespace)
- Generate main wrapper module (just @forward entry point)
- Link: main + accessors + layer + components via iree-link
- Layer (layers/transformer_layer_moe.mlir) now properly tested, not inlined copy

Dependency injection pattern:
- Accessors in @model_params namespace
- Layer imports @model_params.attn_norm_weight etc.
- Dropped get_ prefix - namespace indicates these are parameters
- Cleaner separation: layer declares imports, accessor module provides them

All 47 tests pass, 6 xfailed as expected.
