+++ tests/components/test_attention_block.py
