+++ tests/components/test_embedding_lookup.py
