+++ tests/components/test_moe_ffn_block.py
