+++ tests/components/test_swiglu.py
